refactor(client): components/에 index.ts 추가
