+++ client/src/components/Header/index.ts
@@ -1,0 +1,2 @@
+export { default as BaseHeader } from './BaseHeader';
+export { default as Header } from './Header';
