+++ client/src/components/index.ts
@@ -1,0 +1,6 @@
+export * from './button';
+export * from './Header';
+export * from './input';
+export * from './modal';
+export * from './product-table';
+export * from './toast';
