+++ client/src/components/input/index.ts
@@ -1,0 +1,2 @@
+export { default as Dropdown } from './Dropdown';
+export { default as Input } from './Input';
