+++ client/src/components/modal/LeaveConfirmationModal.tsx
@@ -1,4 +1,4 @@
-import Modal from '@/components/Modal';
+import Modal from '@/components/modal/Modal';
 import { FieldValues, UseFormReset } from 'react-hook-form';
 
 interface LeaveConfirmationModalProps<T extends FieldValues> {
