+++ client/src/components/modal/index.ts
@@ -1,0 +1,2 @@
+export { default as Modal, ModalRow } from './Modal';
+export { default as LeaveConfirmationModal } from './LeaveConfirmationModal';
