+++ client/src/components/product-table/index.ts
@@ -1,0 +1,1 @@
+export { default as ProductTable } from './ProductTable';
