+++ client/src/pages/purchases/list/components/BackorderModal.tsx
@@ -3,9 +3,9 @@ import { useForm, useWatch } from 'react-hook-form';
 import { standardSchemaResolver } from '@hookform/resolvers/standard-schema';
 import { updateBackorderSchema } from '@jumble/shared';
 import { STATUS } from '@/constants/status';
-import Modal from '@/components/Modal';
-import Input from '@/components/Input';
-import LeaveConfirmationModal from '@/components/LeaveConfirmationModal';
+import Modal from '@/components/modal/Modal';
+import Input from '@/components/input/Input';
+import LeaveConfirmationModal from '@/components/modal/LeaveConfirmationModal';
 import UnstyledButton from './UnstyledButton';
 import { useUpdateBackorder, useGetProduct } from '../apis';
 import { useToast } from '@/components/toast';
