+++ client/src/pages/purchases/list/components/PurchaseModal.tsx
@@ -2,12 +2,9 @@ import { useState, useEffect } from 'react';
 import { useForm, useFieldArray } from 'react-hook-form';
 import { standardSchemaResolver } from '@hookform/resolvers/standard-schema';
 import { purchaseSchema } from '@jumble/shared';
-import LeaveConfirmationModal from '@/components/LeaveConfirmationModal';
-import Modal, { ModalRow } from '@/components/Modal';
-import ProductTable from '@/components/ProductTable';
+import { Modal, ModalRow, LeaveConfirmationModal, ProductTable, useToast } from '@/components';
 import { formatDate } from '@/utils/format';
 import { useUpdatePurchase, useGetPurchase, useDeletePurchase } from '../apis';
-import { useToast } from '@/components/toast';
 
 interface PurchaseModalProps {
   purchaseId: string;
