+++ client/src/pages/purchases/list/sections/ResultSection.tsx
@@ -1,8 +1,7 @@
 import { useState } from 'react';
 import { Draft, GetPurchaseResponse, SortBy, SORT_BY, SORT_ORDER } from '@jumble/shared';
+import { Button, Dropdown, useToast } from '@/components';
 import { cn } from '@/utils/cn';
-import Button from '@/components/Button';
-import Dropdown from '@/components/Dropdown';
 import { ValueLabel } from '@/types/value-label';
 import pageLastIcon from '@/assets/page-last-icon.svg';
 import pagePrevIcon from '@/assets/page-prev-icon.svg';
@@ -14,7 +13,6 @@ import PurchaseRow from '../components/PurchaseRow';
 import UnstyledButton from '../components/UnstyledButton';
 import BackorderModal from '../components/BackorderModal';
 import ReceiptModal from '../components/ReceiptModal';
-import { useToast } from '@/components/toast';
 
 interface ResultSectionProps {
   params: Draft;
